[c#] refactor local summary creator into its own util
